Organize repo
This commit is an attempt to make the repo more approachable and clean by reducing component files sizes, creating directories for categories within the hooks dir, and a number of other things.

In addition, this introduces a `Wallet` type which now is never `null`, and instead has a `loaded` boolean property.