Proposal Stats for Live Proposals
Snapshot's GraphQL API does not support accessing data of a live proposal, meaning we cannot view the current results, which are necessary for the forum.

This commit uses Snapshot.js, which offers an API into votes on a proposal, and calculates the state of the proposal from these votes.

Note, we only use this logic IF the proposal is active.  See the `useGetProposalScores` hook for detail.